Rename package to `junit`, rename structs to match
