+++ internal/command/junit/junit_test.go
@@ -1,6 +1,6 @@
 // Copyright (c) HashiCorp, Inc.
 // SPDX-License-Identifier: BUSL-1.1
-package artifact
+package junit
 
 import (
 	"bytes"
